작업 일부  - 클라이언트 - 채팅창 개발  - 클라이언트 - 총알 쏘기/부활/사망 관련 로직 구현  - 클라이언트 ESC키로 그리드 레이어/FPS 라벨 보여줄지 설정가능하도록 함  - 서버 - 총알 쏘기/사망/부활 로직 일부 구현 좀 더 해야함

+++ Projects/TankFighter-Server/Sources/TF/Parser/SendFn.cpp
@@ -59,7 +59,7 @@ void SendFn::SendUpdateFriendListAck(Player* player, int characterUID) {
 
 		// 현재 접속중인지 확인하기 위햇 현재 채널에서 해당 플레이어를 찾는다.
 		Player* pFriendPlayer = pChannel->FindPlayerByCharacterUID(iCharacterUID);
-		info->PlayerState = pFriendPlayer ? static_cast<int>(pFriendPlayer->GetPlayerState()) : static_cast<int>(PlayerState::Disconnected);
+		info->PlayerState = pFriendPlayer ? pFriendPlayer->GetPlayerState() : PlayerState::Disconnected;
 	}
 
 	player->SendAsync(pPacket);
@@ -130,7 +130,7 @@ void SendFn::SendUpdateCharacterInfoAck(Player* player) {
 		CharacterInfo* pMyInfo = &pUpdateCharacterInfoAck->Info;
 		pMyInfo->CharacterUID = spQueryResult->GetInt(0, 0);
 		strcpy_s(pMyInfo->Name, NAME_LEN, spQueryResult->GetString(0, 3).Source());
-		pMyInfo->PlayerState = static_cast<int>(player->GetPlayerState());
+		pMyInfo->PlayerState = player->GetPlayerState();
 		pMyInfo->Win = spQueryResult->GetInt(0, 4);
 		pMyInfo->Lose = spQueryResult->GetInt(0, 5);
 		pMyInfo->Kill = spQueryResult->GetInt(0, 6);
